+++ crates/sqlite-storage/src/models/game.rs
@@ -11,15 +11,15 @@ use serde_json::Value;
 #[diesel(table_name = game_entries)]
 #[diesel(check_for_backend(Sqlite))]
 pub struct GameEntryRow {
-    pub id: i32,
-    pub primary_game_id: i32,
+    pub id: i64,
+    pub primary_game_id: i64,
 }
 
 #[derive(Insertable, Debug)]
 #[diesel(table_name = game_entries)]
 #[diesel(check_for_backend(Sqlite))]
 pub struct NewGameEntryRow {
-    pub primary_game_id: i32,
+    pub primary_game_id: i64,
 }
 
 #[derive(Queryable, Selectable, Identifiable, Associations, Debug)]
@@ -27,8 +27,8 @@ pub struct NewGameEntryRow {
 #[diesel(belongs_to(GameEntryRow, foreign_key = entry_id))]
 #[diesel(check_for_backend(Sqlite))]
 pub struct GameRow {
-    pub id: i32,
-    pub entry_id: i32,
+    pub id: i64,
+    pub entry_id: i64,
     pub name: String,
     pub source_id: String,
     pub lookup_id: String,
@@ -39,7 +39,7 @@ pub struct GameRow {
 #[diesel(table_name = games)]
 #[diesel(check_for_backend(Sqlite))]
 pub struct NewGameRow<'a> {
-    pub entry_id: i32,
+    pub entry_id: i64,
     pub name: &'a str,
     pub source_id: &'a str,
     pub lookup_id: &'a str,
